Emulator prebuilt binaries update (esp. MIPS64-related)
This patch updates the emulator prebuilt binaries to the latest
tip-of-tree from the studio-1.4-dev branch.
The primary motivation for this update was MIPS64 qemu2 updates
and more specifically support for goldfish tty devices;
pair with a kernel that includes commit 24da1af to this end.

See README file for details.

Change-Id: I1a1f084fd6e757b7cfe3f716955d02adc19824eb